Move Avx512bw GetColSums and GetAbsDxColSums to SimdAvx512bwColSums.cpp
Extract Avx512bw::GetColSums and Avx512bw::GetAbsDxColSums from
SimdAvx512bwStatistic.cpp into a dedicated SimdAvx512bwColSums.cpp, and
register the new file in the VS2022 Avx512bw project.
